Recipes list and details page update.
Improve deleting of recipe with confirm window. Get latest recipes in the list. Improve validation of fields when editing or creating a recipe. Update res-api to return needed recipes.

+++ client/src/app/feature/recipes/recipes-detail-page/recipes-detail-page.component.ts
@@ -49,16 +49,18 @@ export class RecipesDetailPageComponent implements OnInit {
     })
   }
 
-  updateRecipe(){
+  updateRecipe() {
     this.makeUpdate = true;
   }
 
 
   deleteRecipeHandler() {
-    console.log('try to delete')
-    this.recipeService.deleteRecipe(this.recipeId)
-      .subscribe((res: any) => {
-        this.router.navigate(['/recipes']);
-      })
+    // console.log('try to delete')
+    if (window.confirm(`Are you sure you want to delete - ${this.recipe.recipeName} ?`)) {
+      this.recipeService.deleteRecipe(this.recipeId)
+        .subscribe((res: any) => {
+          this.router.navigate(['/recipes']);
+        })
+    }
   }
 }
